+++ src/metric-engine/src/engine/create.rs
@@ -13,7 +13,6 @@
 // limitations under the License.
 
 mod extract_new_columns;
-mod validate;
 
 use std::collections::{HashMap, HashSet};
 
@@ -41,9 +40,6 @@ use store_api::region_engine::RegionEngine;
 use store_api::region_request::{AffectedRows, RegionCreateRequest, RegionRequest};
 use store_api::storage::consts::ReservedColumnId;
 use store_api::storage::RegionId;
-use validate::{
-    group_create_logical_regions_requests_by_physical_region_id, parse_physical_region_id,
-};
 
 use crate::engine::create::extract_new_columns::extract_new_columns;
 use crate::engine::options::{set_data_region_options, PhysicalRegionOptions};
@@ -51,8 +47,8 @@ use crate::engine::MetricEngineInner;
 use crate::error::{
     ColumnTypeMismatchSnafu, ConflictRegionOptionSnafu, CreateMitoRegionSnafu,
     InternalColumnOccupiedSnafu, InvalidMetadataSnafu, MissingRegionOptionSnafu,
-    MultipleFieldColumnSnafu, NoFieldColumnSnafu, PhysicalRegionNotFoundSnafu, Result,
-    SerializeColumnMetadataSnafu, UnexpectedRequestSnafu,
+    MultipleFieldColumnSnafu, NoFieldColumnSnafu, ParseRegionIdSnafu, PhysicalRegionNotFoundSnafu,
+    Result, SerializeColumnMetadataSnafu, UnexpectedRequestSnafu,
 };
 use crate::metrics::PHYSICAL_REGION_COUNT;
 use crate::utils::{self, to_data_region_id, to_metadata_region_id};
@@ -96,7 +92,7 @@ impl MetricEngineInner {
                     .await?;
             } else {
                 let grouped_requests =
-                    group_create_logical_regions_requests_by_physical_region_id(requests)?;
+                    group_create_logical_region_requests_by_physical_region_id(requests)?;
                 for (physical_region_id, requests) in grouped_requests {
                     self.create_logical_regions(
                         physical_region_id,
@@ -509,6 +505,39 @@ impl MetricEngineInner {
         };
         [metric_name_col, tsid_col]
     }
+}
+
+/// Groups the create logical region requests by physical region id.
+fn group_create_logical_region_requests_by_physical_region_id(
+    requests: Vec<(RegionId, RegionCreateRequest)>,
+) -> Result<HashMap<RegionId, Vec<(RegionId, RegionCreateRequest)>>> {
+    let mut result = HashMap::with_capacity(requests.len());
+    for (region_id, request) in requests {
+        let physical_region_id = parse_physical_region_id(&request)?;
+        result
+            .entry(physical_region_id)
+            .or_insert_with(Vec::new)
+            .push((region_id, request));
+    }
+
+    Ok(result)
+}
+
+/// Parses the physical region id from the request.
+fn parse_physical_region_id(request: &RegionCreateRequest) -> Result<RegionId> {
+    let physical_region_id_raw = request
+        .options
+        .get(LOGICAL_TABLE_METADATA_KEY)
+        .ok_or(MissingRegionOptionSnafu {}.build())?;
+
+    let physical_region_id: RegionId = physical_region_id_raw
+        .parse::<u64>()
+        .with_context(|_| ParseRegionIdSnafu {
+            raw: physical_region_id_raw,
+        })?
+        .into();
+
+    Ok(physical_region_id)
 }
 
 /// Creates the region options for metadata region in metric engine.
